use generic molgraph class
